// packages/monorepo/src/nx-types.ts
export namespace Nx {
// ...
  export interface RunManyOptions {
    /** Task to run for affected projects */
    readonly target: string;
    /** This is the configuration to use when performing tasks on projects */
    readonly configuration?: string;
    /** Exclude certain projects from being processed */
    readonly exclude?: string;
    /**
     * Do not stop command execution after the first failed task.
     */
    readonly noBail?: boolean;
    /**
     * Defines how Nx emits outputs tasks logs
     * @default "stream"
     */
    readonly outputStyle?:
      | "dynamic"
      | "static"
      | "stream"
      | "stream-without-prefixes";
    /**
     * Max number of parallel processes
     * @default 3
     */
    readonly parallel?: number;
    /** Project to run as list project names and/or patterns. */
    readonly projects?: string[];
    /** This is the name of the tasks runner configuration in nx.json */
    readonly runner?: string;
    /** Rerun the tasks even when the results are available in the cache. */
    readonly skipCache?: boolean;
    /** Ignore cycles in the task graph */
    readonly ignoreCycles?: boolean;
    /** Prints additional information about the commands (e.g. stack traces). */
    readonly verbose?: boolean;
  }
}